fragsifier/requirements.txt
pandas==0.22.0
pathlib2==2.3.0
numpy==1.13.3
tqdm==4.19.4
termcolor==1.1.0
biopython==1.69
scikit_learn==0.19.1
scipy==1.0.0
termcolor==1.1.0 
